YE-2707 Movies to support more then one set/collection
By default kodi only supports one set per movie but we need to have movies in multiple sets.
After digging into the internals to try to make a "upstreamable" implementation I decided
to do just a quick implementation that fits our needs, as it would take far too long at this stage
to modify the default handling in kodi.

This implementation now just adds the new table for storing the links betweens movies and sets
and then "unwraps" that list to make it work with the current handling in kodi.