Game add-ons: Fix segfault for some cores
The cause of the segfault is libretro cores that fail to allocate input
resources during initialization. Instead, input resources are allocated
when the game is loaded. To accomodate this, we move our controller
connecting/disconnecting to after the game is loaded.

This commit also corrects the order of init/deinit for other game add-on
resources.