[YC] FIXED reloading of pre-loaders
It was happening that when navigating back to the home screen the pre-loaders where triggered again.
Normally this does not cause any issues, but if there are too many items in the db or the connection
is slow, it was the case the initializing indicator was shown in the UI. As we do not need to reload
any of the smartplaylists I added to ignore any force-reloads of playlists.