VideoPlayer: Rendering - replace gl ARB functions with std ones
All the ARB postfixed GL functions we use are already
part of std GL.
ARB symbols are not guaranteed to be exposed by the ABI,
which results in undefined symbols when you link against
libOpenGL.so instead of libGL.so.

With this change we can link against libOpenGL.so successfully.