Update README.md to list BSD as its own operating systems
BSD (Berkeley Software Distribution) based operating systems like example FreeBSD are not just another Linux distribution so "BSD" should reasonably be listed as its own operating syste. There are a lot of similarities and OS compatibility layers with Linux, but those does not change the fact that BSD is not Linux.