AEStreamParser: Fix duplicated frame when syncFunc requests more data
CAEStreamParser::AddData() gets called with a bufferSize in-out
parameter that represents the output buffer size on input and actual
data length on output.

However, it is not set to zero when the sync function has requested more
data and no frame was actually output, causing the caller to reuse
whatever data was in the buffer (usually a previous frame).

Currently only the DTS sync function uses this functionality. It
requests more data when the frame is exceptionally short (<13 bytes) or
when there is not enough data to check for DTS HD presence.

Fix the issue by adding the missing zeroing.