DVDVideoCodecAndroidMediaCodec: Revert c15ff2d9f6e
This restores acceleration for DVDs. It was originally disabled, because someone had issues with something (not documented). As no one wanted to track down the real issue - a quick hack was chosen. Since then users screamed on regularly basis until it was reenabled again. Let's see which errors come up this time.