[YC] Fix Audio/Subtitle InfoLabel values
As we want to show the available audio/subtitle
streams before we start the movie, we must not
rely on the actual file stream, but on the infos
stored in the database.
This is only done for ListItem.* infolabels, not the
VideoPlayer.* labels.

Furthermore we are interested in the languages and don't
care how many english streams we have, that's
why we return a distinct list.

Ticket: YE-3524