[YC] Do not execute count query if not needed
For some reason the count query of movies is substantially
slower than the query to acutallly retrieve the movies,
although they use the same where clauses etc.
It's roughly 1s vs. 7s.

When we do not set limits the default is 0/-1.
So we don't want to execute the count statement in
this case at all.

Ticket: YE-4678, YE-4592