[YC] Fixed random results on movie sets
Due to the implementation of multiple sets per movie, this caused
that the assigned set of a movie could be different. Now instead
we are creating a clone of the movie and assigning that set, so
we are sure that the correct set is used.

This can be rebased onto
YE-2707 Movies to support more then one set/collection
later on.