[addons] add define ATTRIBUTE_FORCEINLINE on addon header
This add a OS independent defination to force a inline on a function.

This is required on some screensaver and visualization addons to have a clean
code with better performance.

This change does not affect the "C" interface and all previous compiled addons
still work.

(cherry picked from commit 94c100548c92849077463522df492298af59fc20)