Fix filter performance
Our movie odb query's join conditions
need to be defined at compile time, that
leads to a big issue when using smart filters, as
too many tables are joined that we do not even
need, hence some statements get damn slow (like 30s).

In order to fix this, we now use the original smart
playlist filter implementation, where all filters
can be expressed as subselects in the where clause, thus
we don't need to define any joins.
This where clause can be fed into odb at runtime as a string.

To achieve this I needed to fix a lot of the smart playlist
code that we didn't touch at all, so a lot was referencing
old fields and tables that do not exist anymore.