Make file a lazyptr again
Due to CODBFile not being part of the
foreign section anymore, we always hit the
database when iterating over the movie result
set.
We can already query all needed fields in the
original view and thus have all infos.
This can therefore safely be in a section again
and ideally never needs to be loaded.