Always read added date for tvshows
We need to join all the way down to the files
of episodes in order to get the date added.
This is needed for the sort.

If you only grab date added, when getDetails is set, the
sort order changes after a user enters into the season
view and returns from it.