Fix grabbing of sets
The GetSetsNav method relies on
the m_sets info of the movies'
VideoInfoTag.
This is normally only filled during
GetDetailsForMovie as it is expensive.
But we usually haven't called this method for
all movies yet.

So what we do is:
Select all set/movie correlations and store them
in a map when GetSetsByWhere is called.
After getting the movies, iterate through the item
collection and add the setinfo for each item.