Improve speed of cover loading
Our background loaders call GetArtForItem with
a single item as parameter.
This results in more than 3 sql statements per item
and guess what happens if you have 10k items.

Now when artwork for an item of a type (eg. movies)
is requested, we'll cache all artwork for the type and
return the art for the requested item.
This works within 1 sql statement.

Whenever the next artwork is requested we can deliver it
from the cache and no further db hits need to happen.