Fix album sorting via db
Unfortunately it's impossible to have a parameterized
where clause, a fixed GROUP BY and a dynamically created
sort clause in odb.
Therefore we need to add the album group by to the places
where it's queried, so that the "group by" ends up between
the "where" and the "order by".