Remove on the fly db lookup of translations
This leads to slow cover loading, because the optimal
case that every single art is cached, does not seem to
be true in the real case scenario.
For now we don't lookup on the fly.

Normally all translations are cached when a language change
happens or the player boots, so if we miss translations
it's also highly possible that we don't find them anyway.