Fix filter being applied on child items
This was a particular problem for playlists.
Albums worked by chance.

Prior to this change if you filtered playlists
by title, the title filter was also applied to songs
when you entered the playlist.
In some situations your playlist was shown as empty then.

Of course this is not wanted.
For Yachtcloud we have no use case where we want to apply
filters for subfolders, as our skin only has 2 layers per
content type.