Fix genres in filter are not translated
We need to translate the genres when we show
the filter dialog.
This is a bit tricky, because as of now, we use
the genre names in the where clause.

In order to fix the issue following steps are done:
 - Translate genres when we grab them from the db
 - When we build the smart playlist filter we translate
   the strings to the corresponding ID's
 - When you open the filters again, we need to translate
   the ID's into there translated names again

Above is done by querying into the translation table.

Ticket: YE-6181