Fix album path when entering from artist search
When you search for a term, hit artists as a result and
enter the artists, you will be displayed with albums of this
artist.

Without the m_bIsFolder->true this will show item ID's instead of
album names and the select action is attempting to directly play.

Additionally the path was built wrongly, because when you enter an
artist the URL is musicdb://artists/875 and if we append the album ID
we end up with a path like: musicdb://artists/875/234 which seems
utterly wrong.

To fix it, we always make sure that album items have a path like:
musicdb://albums/123/.

Ticket: YE-6076