[YC] Reload translation on cache misses
If we miss a translation in the cache, we try once
to reload the whole translation table.
Within one db load, this is only done once because sometimes
there is missing data in translations and we don't want to
reload on every translation process.

When we call one of the GetXXXXNav-methods again and have a
cache miss, we'll try to reload the table ONCE again.

Ticket: YE-6170