Fix short UI freeze when entering main sections
In order to have a flawless GUI experience, we
want to load movie, tvshow, music, playlist views
in the background and bring a spinner if it takes
long.

In order to have this working we introduce a boolean
condition for containers that indicates if loading
has finished.
This allows the skin to show "no content" messages after
the loading has finished.