Fix media window path history
Usecase:
You enter the playlist screen but decide to abort
when the spinner is shown.
Now you endup in the filedirectory.

Reason:
Getting the previous path seems to be implemented
in a weird way, it's always the current directory and
as we do not want to endlessly load the same content
we need to fix the previous path retrieval.

Fix:
Do not clear the window history.
Get the previous used path from our path history instead
of using m_vecItems->GetPath().