VideoPlayer: Make CloseFile asynchronous
When doing start/stop actions rapidly via JSON
calls and the user has refreshrate switch enabled
there is a high chance of a deadlock becasue
we block the mainthread in CloseFile but also want
to switch refreshrate from the VideoPlayer thread.

Don't wait for the VideoPlayer thread to exit in
CloseFile and destroy the player in PlaybackCleanup
fixes the problem.