InputManager: Fix JSON ButtonEvent not handled if it has longpress mapping
As the XBMC_BUTTON event is handled by using OnKey, the CKey object
needs to have FromService = true in order to get handled.
That's because usually OnKey is followed by OnKeyUp and only if this
combination is used, keys with a longpress mapping are handled.
Otherwise it fails to resolve and nothing happens at all.